+++ database/migrations/2021_01_04_200910_create_order_dishserving_table.php
@@ -15,9 +15,9 @@ class CreateOrderDishservingTable extends Migration
     {
         Schema::create('order_dishserving', function (Blueprint $table) {
             $table->unsignedBigInteger('order_id');
-            $table->unsignedBigInteger('dish_id');
-            $table->unsignedBigInteger('serving_id');
-            $table->unsignedInteger('count');
+            $table->unsignedInteger('dish_id');
+            $table->unsignedInteger('serving_id');
+            $table->unsignedTinyInteger('count');
             $table->foreign('order_id')->references('id')->on('orders');
             $table->foreign(['dish_id', 'serving_id'])->references(['dish_id', 'serving_id'])->on('dish_serving');
             $table->primary(['order_id', 'dish_id', 'serving_id']);
